+++ ebustoolbox/schedule_readers.py
@@ -6,6 +6,7 @@ from pathlib import Path
 from typing import Callable, Type
 from abc import ABC, abstractmethod
 from uuid import UUID
+from tqdm.auto import tqdm
 
 import eflips
 from django import forms
@@ -14,6 +15,7 @@ from inspect import signature
 
 from eflips.ingest import DummyIngester, AbstractIngester
 from eflips.ingest.dummy import BusType
+from eflips.ingest.vdv import VdvIngester
 from sqlalchemy import create_engine
 from sqlalchemy.orm import Session
 
@@ -37,6 +39,8 @@ def get_options_form(reader_num: int):
             return SimbaScheduleReader.get_options_form()
         case 2:
             return EflipsIngestScheduleReaderDummy.get_options_form(DummyIngester)
+        case 3:
+            return EflipsIngestScheduleReaderVDV.get_options_form(VdvIngester)
     raise NotImplementedError
 
 
@@ -106,6 +110,9 @@ def get_schedule_reader_factory(reader_num: int) -> type(ScheduleReader):
             return SimbaScheduleReader
         case 2:
             return EflipsIngestScheduleReaderDummy
+        case 3:
+            return EflipsIngestScheduleReaderVDV
+
     raise NotImplementedError(f"Schedule Reader with {reader_num} not found")
 
 
@@ -206,17 +213,24 @@ class SimbaScheduleReader(ScheduleReader):
         route_id = 1 if Route.objects.last() is None else Route.objects.last().id + 1
         trip_id = 1 if Trip.objects.last() is None else Trip.objects.last().id + 1
         line_id = 1 if Line.objects.last() is None else Line.objects.last().id + 1
-        for rotation_id, rotation_trips in trip_data.items():
+        for rotation_id, rotation_trips in tqdm(trip_data.items()):
             sorted_trips = sorted(rotation_trips, key=lambda trip: trip["departure_time"])
             prev_arrival_time = sorted_trips[0]["departure_time"] - timedelta(hours=1)
-            for trip in rotation_trips:
+            prev_arrival_name = sorted_trips[0]["departure_name"]
+            for trip in sorted_trips:
                 if trip["departure_time"] < prev_arrival_time:
                     self.errors.append(
                         f"The following trip overlaps with another. {trip} departs "
                         f"before the previous arrival at {prev_arrival_time}"
                     )
                     raise self.SimbaScheduleReaderException
+                if trip["departure_name"] != prev_arrival_name:
+                    self.errors.append(
+                        f"The following trip does not end at the previous arrival station {prev_arrival_name}: {trip} "
+                    )
+                    raise self.SimbaScheduleReaderException
                 prev_arrival_time = trip["arrival_time"]
+                prev_arrival_name = trip["arrival_name"]
                 if trip[self.LINE] not in line_dict:
                     line = Line(scenario=scenario, name=trip[self.LINE], id=line_id)
                     line_id += 1
@@ -302,6 +316,7 @@ class SimbaScheduleReader(ScheduleReader):
             default_params = {
                 "scenario": scenario,
                 "name": name,
+                "name_short": name,
                 "battery_capacity": self.default_capacity,
                 "charging_curve": [[0, self.default_capacity], [1, self.default_capacity]],
             }
@@ -633,5 +648,26 @@ class EflipsIngestScheduleReaderDummy(EflipsIngestScheduleReaderBase):
             "rotation_per_line": rotation_per_line,
             "opportunity_charging": opportunity_charging,
             "bus_type": bus_type,
+            "progress_callback": None,
+        }
+
+
+class EflipsIngestScheduleReaderVDV(EflipsIngestScheduleReaderBase):
+    """
+    This class is a dummy shim for the eflips-ingest DummyIngester class. It is used for testing the shim system.
+    """
+
+    def __init__(
+        self,
+        x10_zip_file: str,
+    ):
+        super().__init__()
+        self._ingester = VdvIngester(self._database_url)
+
+        # random_text_file is a Path as string, we need to convert it to a Path
+        x10_zip_file = Path(x10_zip_file)
+
+        self._kwargs = {
+            "x10_zip_file": x10_zip_file,
             "progress_callback": None,
         }