+++ elevation_api/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class ElevationApiConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutourls.pyField"
+    name = "elevation_api"